feat: create component to activity flat list
